separate file: hook.c
