+++ Makefile
@@ -32,6 +32,7 @@ SRCS		:=	cylinder/cylinder_color.c \
 				sphere/sphere_distance.c \
 				display/init.c \
 				display/display.c \
+				display/hook.c \
 				display/screen.c \
 				display/set_each_pixel.c \
 				display/ray_trace.c \
